Always replicate rtn submission data when missing (#1058)
https://eaflood.atlassian.net/browse/WATER-4855

> Part of the work to migrate management of return requirements from NALD to WRLS

We've realised NALD has return log data that we don't. Initially, we thought it was just the return lines for pre-2013 return logs. But after some digging, it appears there are not only the return submission lines but also some return logs we don't have that NALD does.

Putting aside the missing return logs, we need to get those missing return submission lines in. We have found that NALD structures the return logs and submission data very differently from WRLS. Our plan to do something 'quick & dirty' with a SQL query is therefore not possible. That means we'll need to code a solution.

We intend to stop all 'return leg' processes shortly, just as soon as WRLS takes over from NALD on all things 'returns' related. So, it makes no sense to start from scratch.

In recent changes, we built on the existing logic, moving [Moving the import of return logs into their own job](https://github.com/DEFRA/water-abstraction-import/pull/1055), [Refactoring it to use DB rather than API requests](https://github.com/DEFRA/water-abstraction-import/pull/1056) and [Replacing the faulty replicate returns logic](https://github.com/DEFRA/water-abstraction-import/pull/1057).

The next step is to take the now fixed replicate logic and use it wherever we find that a completed return log is missing submission data.

We're working with the assumption that if submission data already exists, it was either because we have previously replicated it from NALD or because it was entered in WRLS directly (so it takes precedence). With this change in place, we can start to bring in the missing data (at least for 2013 onwards).