Include old return lines in return logs import
https://eaflood.atlassian.net/browse/WATER-4855

> Part of the work to migrate management of return requirements from NALD to WRLS

We've realised NALD has return log data that we don't. These are turn return lines dated pre-2013.

When the NALD import was first put together it was found that the FME job was unable to complete creating the NALD extract zip file overnight if the NALD return lines table was included. A decision was made to filter out those dated prior to 2013, specifically `WHERE (ARFL_DATE_FROM >= TO_DATE('20130101','YYYYMMDD'))`.

Filtering these out meant the volume of data was reduced enough that the FME job could complete generating in the extract in less than an hour, giving the imports in this project enough time to complete overnight.

Shame, because it turns out we've never imported this line data! 🤦😂

In our recent changes to get the import ready for importing it, we've found that the legacy [water-abstraction-ui](https://github.com/DEFRA/water-abstraction-ui) and [water-abstraction-service](https://github.com/DEFRA/water-abstraction-service) pulls return submission data older than '2018-10-31' directly from the `import` schema. So, we only have a few records in WRLS itself that are dated prior to 2018.

You can't make this stuff up!

Anyway, we now intend to bring in _all_ the missing data. The change [Always replicate return submission data when missing](https://github.com/DEFRA/water-abstraction-import/pull/1058) updated the return log import to always attempt to add the missing submission data where it does not currently exist. But it is only looking at whatever is found in `NALD_RET_LINES` in the nightly NALD extract zip file.

We have obtained a copy of all the pre-2013 NALD return lines. We're still working on how we make them available in our AWS environments, but we know they will be in `public.NALD_RET_LINES`. So, this change updates the 'return-logs' import to

- check if the table exists
- if it does, run an extra query for those return logs that start prior to 2013 to include their return lines when attempting to replicate the submission data