Add old return lines exists flag
We have the queue job check to see if the table we'll be holding the old NALD return lines extract in exists. We do it in the queue so we only fire the query once.

We add the result as a flag to be passed to each of the import jobs.