Update replicate returns to check the old lines table
We add logic to `replicate-returns` to check the old return lines table when the flag states it exists.

We add some logic to avoid querying the table unnecessarily, as we know which lines exist in it and which don't.

Unfortunately, the extracts don't appear to be exactly the same (the date formats are different). So, rather than mess too much with the data we simply just fire 2 versions of the query at the two tables and combine the results.