Pause to go look at the return logs code
This is working, in so far as it is finding the lines, combining the data, and persisting without error to the DB.

The problem is we came across a problem replication on our second example licence.

`7/34/13/*G/0299A` has return lines that match to a return log the legacy import code is generating. The period is 2008-04-01 to 2009-03-31. Our ned code is finding the lines perfectly fine, but we're not matching them to the blanks as expected.

After investigating, it is because the return lines have been captured as monthly, but the row we're being passed from the legacy import code has the return log as daily. Hence the blank lines we are generating are not matching to the return lines the query is finding.

So, we're pausing on this work whilst we go and take a long overdue look at the return log generation logic.