Fix issue
We were wrong. It wasn't that the generated return log had got it wrong, and was marked as daily when it should have been monthly. The way the data had been submitted made it _look_ like a monthly return, when in fact it is daily.

The reason the lines were not matching was because we were not ordering the old lines in the same was as the post-2013 lines. Because of the differing date formats, `RET_DATE` for the old returns when treated as a string is not ordered correctly. And we have logic that bails the total quantity loop when it sees the blank line's end date go past the NALD line's. The lines being out of order meant we were bailing before it had a chance to match the correct lines.