Download, extract and load old NALD return lines
https://eaflood.atlassian.net/browse/WATER-4855

> Part of the work to migrate management of return requirements from NALD to WRLS

We've realised NALD has return log data that we don't. These are turn return lines dated pre-2013.

When the NALD import was first put together, it was found that the FME job could not complete creating the NALD extract zip file overnight if the NALD return lines table was included. A decision was made to filter out those dated before 2013, specifically `WHERE (ARFL_DATE_FROM >= TO_DATE('20130101','YYYYMMDD'))`.

Filtering these out meant the volume of data was reduced enough that the FME job could generate the extract in less than an hour, giving the imports in this project enough time to complete overnight.

We've now been provided a one off extract of these missing old return lines, and have [updated the return logs job to include them](https://github.com/DEFRA/water-abstraction-import/pull/1059).

But for this to happen, we need to get the data into a table in the DB. We've placed the extract in the same AWS bucket we extract the NALD nightly import from. This change uses the same functionality to download, and extract the data (a CSV file), then loads it into a new table we'll add to the `public` schema.

It will only do this once per environment, though we'll need to check each time the job is run.