Amend return-logs import to exclude fortnightly
https://eaflood.atlassian.net/browse/WATER-4856

> Part of the work to migrate management of return requirements from NALD to WRLS

As part of getting ready for WRLS taking over management of returns from NALD, we looked into whether there were any discrepancies in the returns data.

We found two things.

- What was originally just though to be missing line data for returns older than 2013-01-01 turned out to be nearly all submission data older than 2018-10-31
- There are 338 NALD form logs (return logs) linked to return formats (return requirements) that don't exist in WRLS

The last set of changes to this app were all in support of the first issues, and the aim of getting the missing submission data in. This changes relates to the second.

We have found that in nearly all cases, there is a valid reason for why the 338 have not been imported. In essence, the licence has ended, and where the end date falls means that the form logs no longer intersect with valid return periods.

For example, licence `01/123` has a return format which starts on 1998-11-16. The licence was revoked 1999-03-01. `src/modules/return-logs/lib/transform-returns.js` determines just one return period which is based on the winter or summer return cycles, the start and end times of the format, and finally the licence.

```javascript
  cycles: [
    { startDate: '1998-11-16', endDate: '1999-03-31', isCurrent: true }
  ]
```

There are a number of form logs for the licence, but the first starts on 1999-04-01. It nor the rest intersect with return period the import engine has calculated, so it does not get imported.

The one exception is a licence that has a return format where the reporting frequency has been set to `fortnightly`. WRLS does not recognise this frequency in its `returns.returns` table, so this licence was erroring on import.

This change is to better handle the error.