Fix incorrect status on later return logs
Spotted because of the 'less than 2018' clause in the update, the process was adding submissions for returns but leaving their status as `due` and their received date blank.

On reflection, there is no reason why there should be a difference. Though submissions are now entered in WRLS, NALD is still classed as the `main` record. So, there is no reason not to update in the same way no matter the records year.