Update Nil return handling
We've realised that NALD is storing 0 for lines within the abstraction period for a return log, irrespective of whether someone submitted something or not.

For example, a weekly return log with a full return period, and an abstraction period between 1 May and 30 June, will have 365 lines (because NALD holds everything at the daily level). Then between May and June, every 7th line will have a 0.

This doesn't mean someone submitted a zero, its just that NALD defaults to it. We'd assumed all lines would be null for a Nil return.

So, knowing this, we can streamline our returns replication to ignore all null and 0 lines. With that change to the query, if we get nothing back, we have a Nil return. If we do, it must be a quantity, so we can stop worrying about `0` and `null`.