Delete redundant tests and move possible keepers
We've made a large swathe of the tests redundant with our changes. The rest were pointless, for example, that a job was registered rather than what the job was doing or a query was called instead of that the query was right.

So, we're dropping them.

Any that we think might have value, we've moved into the processes we think they are most relevant to. We'll take a look proper look and fix them up next.