Migrate crm v2 transformer company tests
We first move the data fixture they'd create to a support folder, and add a new method for role.

We then migrate the 'essence' of the existing tests. They old ones focused on invoice accounts but the import of this was disabled many years ago. So, we dropped all that pointless work and code in these changes.