Delete return-helpers tests
These tests don't align with what we actually have in `return-helpers.js` any more. That is the first reason for deleting them.

But also after checking them they are simply asserting a particular query is called when a helper is called.

So, there was no value in the tests in the first place. Therefore, we've opted to just delete the file.