Migrate Nald test
The original test was linked to the original `nald-import` job. It both extracted the NALD data zip and imported licences to the `permit` schema.

As such, the test kind of covered both things.

We've split the old job into two distinct processes. So, likewise, we've split what the old NALD test was doing into two.