Update onComplete handlers
We make two changes.

- we use the word 'completed' rather than 'finished' as this ties in with the vernacular of pg-boss
- we always trigger the next job rather than stop processing

After review, each job is distinct and not 100% dependent on the previous (it might not have the most up to date data to work with but it will still run).

Therefore, there is no need to stop processing of the jobs should one fail. We can continue to the next.