Reformat job completion email
Realised that the current content not only required an overly complex query, it didn't really tell us anything.

With the reformat now we can see each step that was processed, when it was started, and how long it took.

It might also help those who receive the email get a better sense of what the overnight import is actually doing.