Add message handling to the processes
What would be great is if the email that gets sent at the end of the job queue could tell us more than pass/fail.

We're going to update the content to pull the list of steps and their timings from the jobs table. But it would also be great if we could add to each step any useful messages from the processes.

This change enables that. For example, now if we skip a step because of some config setting, we can record that in the email. Another example is noting licences without start dates.

We also ensure any errors thrown will be included in the email. We can then use this trigger to check the logs for a more detailed reason for the failure.