Refactor all overnight jobs into one (#1063)
https://eaflood.atlassian.net/browse/WATER-4855

> Part of the work to migrate management of return requirements from NALD to WRLS

This will hopefully be the final change in a series.

During the work to migrate returns management from NALD to WRLS, we realised NALD has return submission data that we don't. These are return lines dated pre-2013.

When the NALD import was first put together, it was found that the FME job could not complete creating the NALD extract zip file overnight if the whole NALD return lines table was included. A decision was made to filter out those dated before 2013, specifically `WHERE (ARFL_DATE_FROM >= TO_DATE('20130101','YYYYMMDD'))`.

We later discovered that when return submissions were taken over by WRLS in 2018, no NALD submission data was imported. It appears to exist when you look at post-2013 returns in the UI. But behind the scenes, the [water-abstraction-service](https://github.com/DEFRA/water-abstraction-service) is dipping into the `import` schema's tables for any return log before 2018-10-31.

In the following changes, we've updated what the legacy import process was doing with returns to ensure this missing submission data will be imported and ready for WRLS to take over returns completely.

- [Stop 'return leg' processes with feature flag](https://github.com/DEFRA/water-abstraction-import/pull/1054)
- [Move import of return logs into their own job](https://github.com/DEFRA/water-abstraction-import/pull/1055)
- [Refactor new return-logs import job to use DB](https://github.com/DEFRA/water-abstraction-import/pull/1056)
- [Replace faulty replicate returns logic](https://github.com/DEFRA/water-abstraction-import/pull/1057)
- [Always replicate rtn submission data when missing](https://github.com/DEFRA/water-abstraction-import/pull/1058)
- [Include old return lines in return logs import](https://github.com/DEFRA/water-abstraction-import/pull/1059)
- [Download, extract and load old NALD return lines](https://github.com/DEFRA/water-abstraction-import/pull/1060)
- [Amend return-logs import to exclude fortnightly](https://github.com/DEFRA/water-abstraction-import/pull/1061)

The problem is that it is another overnight process which queues up the same 74K licences. Several other jobs also queue up licences for processing. Though only in our local environments, we've seen that it takes hours to clear a 74K job queue where the job is 'do nothing'.

It's time to refactor the import process to behave as everyone automatically assumes it is: one overnight import process.

This simplifies the app and means we are only queuing the licences to be processed once. Just doing that should speed things up. However, we can also use asynchronous Javascript and process multiple 'processes' concurrently for the same licence. This change will give us more confidence that the import _can_ be completed each night, especially as we are about to add more load.

So, this change

- Will refactor any logic in the current 'jobs' into 'processes'
- The existing API triggers will be refactored to use the 'processes' instead of jobs (which will be much easier for testing/investigation!)
- A new 'import-job' will be created to be the app's only 'true job'. This will call on the 'processes'

Finally, bye-bye [pg-boss](https://github.com/timgit/pg-boss)!

We've got no issues against it. It was just unnecessary for what was being done here, and removing it gives us back some resources and allows us to delete a bunch of code, whilst making the whole process simpler and easier to follow.