Update return log frequency
First pass of testing realised that we had inadvertently applied the existing NALD frequencies of quarterly to our return logs, even though we only support daily, weekly and monthly.

Then realised we might have done this for real. So, to ensure we have the frequencies set only to those we support before we turn things off we update the 'update' command to also overwrite frequency.