Add matched property to nald lines
We'll use this to solve our problem of how to use these NALD lines that are for frequencies we don't support.

By flagging a NALD line as matched, we know we'll only apply its quantity once in the return submission. This avoids the issue we faced when we first spotted these unsupported frequencies of line quantities being applied multiple times because the NALD periods matched to multiple WRLS blank lines.