Handle importing non-WRLS frequency rtns from NALD (#1072)
https://eaflood.atlassian.net/browse/WATER-4855

> Part of the work to migrate management of return requirements from NALD to WRLS

In the following changes, we updated the legacy import process's handling of returns to ensure the missing submission data was being imported ready for WRLS to take over.

- [Stop 'return leg' processes with feature flag](https://github.com/DEFRA/water-abstraction-import/pull/1054)
- [Move import of return logs into their own job](https://github.com/DEFRA/water-abstraction-import/pull/1055)
- [Refactor new return-logs import job to use DB](https://github.com/DEFRA/water-abstraction-import/pull/1056)
- [Replace faulty replicate returns logic](https://github.com/DEFRA/water-abstraction-import/pull/1057)
- [Always replicate rtn submission data when missing](https://github.com/DEFRA/water-abstraction-import/pull/1058)
- [Include old return lines in return logs import](https://github.com/DEFRA/water-abstraction-import/pull/1059)
- [Download, extract and load old NALD return lines](https://github.com/DEFRA/water-abstraction-import/pull/1060)
- [Amend return-logs import to exclude fortnightly](https://github.com/DEFRA/water-abstraction-import/pull/1061)

Mostly!

When we realised this was going to slow down the import too much, we also made [a big change](https://github.com/DEFRA/water-abstraction-import/pull/1063) to address a number of long-standing issues with the app.

While doing that, we spotted that NALD supports return frequencies that WRLS doesn't: fortnightly, quarterly, and yearly. To get things moving and focus on the refactor, we had the returns import process ignore submissions for these.

Now it is time to get these last remaining returns importing.

The weirdness in NALD with these lines is that they have funny periods, for example, `2004-03-01 to 2004-06-30` (there doesn't appear to be any consistency or correlation with the idea of fortnightly, quarterly or yearly 🤯 😩 ).

In our existing engine, the line would match to multiple WRLS lines we'd generated, leading to a duplication in quantities applied.

We realised we could solve this by adding a `matched` flag to the NALD lines we retrieve. If `replicate-returns.js` matches the NALD line to a WRLS blank, it flags the NALD line as matched. This means it won't be used again when determining the total for other WRLS lines. Now, the quantity is only applied once.