Skip cleaning return data if import disabled
https://eaflood.atlassian.net/browse/WATER-4890

> Part of the work to migrate management of return requirements from NALD to WRLS

We have completed the work to take over the management of return versions and return requirements (though not yet live). We are just completing all the work to manage the creation and reissuing of return logs based on them.

When both go live in WRLS, the 'returns leg' will become defunct in NALD. This means we must stop importing return version, requirement, and return log data. All changes are required to be done in WRLS, so anything done in NALD will be considered a mistake and ignored.

We [added a feature flag to stop the import](https://github.com/DEFRA/water-abstraction-import/pull/1054). But when we did our [big refactor](https://github.com/DEFRA/water-abstraction-import/pull/1063) we split the return version data cleaning and import tasks.

And we've just spotted that the cleaning is no longer behind this flag. This means if we were to disable the import and WRLS took over, the clean step would delete anything created directly in WRLS because it would find no match in NALD.