Fix licence-crm-import
https://eaflood.atlassian.net/browse/WATER-4855

> Part of the work to migrate management of return requirements from NALD to WRLS

After [refactoring the import into a single job](https://github.com/DEFRA/water-abstraction-import/pull/1063) when we tested it in `tst` and `pre` we found the licence-drm-import failed for a number of records.

Looking into the error, it seems folks have gone a bit crazy with the indexes and constraints on this table, so though it worked fine locally our UPSERT was failing where a document header record had a verification ID.

Switching the constraint to be `system_external_id` appears to make it work in all environments.

Go figure!

We also spotted that we were reusing a numbered param. Handily, it seemed to be working fine. But normally we repeat value not the param (e.g. $2) so we corrected that as well.