Remove auth from test endpoints
https://eaflood.atlassian.net/browse/WATER-4855

As part of [refactoring the import into a single job](https://github.com/DEFRA/water-abstraction-import/pull/1063), we made it possible to trigger the individual steps directly.

The app has auth, but it is simply checking the request's auth header against a value stored in plain text in the server. So, its not really adding any protection, and the steps themselves are not destructive.

So, to allow us to more easily debug issues in our AWS environments we are marking them all as no-auth.