Wrap individual licence data import in try/catch (#1078)
https://eaflood.atlassian.net/browse/WATER-4855

> Part of the work to migrate management of return requirements from NALD to WRLS

When we did the [Refactor all overnight jobs into one](https://github.com/DEFRA/water-abstraction-import/pull/1063) we also overhauled the way we communicate what happened during the import via the completion email.

We also needed a solution that would allow us to run certain licence import tasks synchronously, whilst ignoring returns if they have been disabled.

Those changes meant that the licence data import task was being done outside of a try/catch. Specifically, `LicenceCrmImportProcess` wasn't included in the `Promise.allSettled` that occurs.

We became concerned that this might result in an error preventing the processing of the other licences. To protect against this, we moved the processing of each licence into its own function wrapped in a try/catch to allow the processing of other licences to continue.