Bring omfg() in sync with water-abstraction-system (#1079)
https://eaflood.atlassian.net/browse/WATER-4855

> Part of the work to migrate management of return requirements from NALD to WRLS

While working on the import service, we spotted that the base notifier in this repo is out of sync with the one we copied it from in [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system). Most notably, the order of arguments to `omfg()` is out of sync.

This updates it to be in sync.