Extract NALD data directly into import schema
https://eaflood.atlassian.net/browse/WATER-4855

> Part of the work to migrate management of return requirements from NALD to WRLS

Whilst testing getting the missing returns data into our non-prod environments we hot an issue of the DB hitting the allocated size.

We realised one of the reasons we need to allocate more space than we actually use is because the extract nald data process first loads everything into `import_test`. Once loaded it then swaps the schemas.

But this means during the load we are duplicating the import schema, which is why we need to much extra allocation space. If we dropped the import schema, and then loaded the extract CSV files straight in, we'd avoid duplicating anything and get that 'cushion' back.

We understand why it was done. Because of the decision to fetch returns data directly from the import schema when viewing the in the legacy UI, the previous team needed to add in this precaution in case the import failed.

But as we are now moving to our new view return which will be based solely on data in WRLS, this caution is no longer necessary.