Extract NALD data directly into import schema (#1080)
https://eaflood.atlassian.net/browse/WATER-4855

> Part of the work to migrate management of return requirements from NALD to WRLS

Whilst testing getting the missing returns data into our non-prod environments, we encountered an issue with the DB hitting its allocated size.

We realised one of the reasons we need to allocate more space than we actually use is that the extract nald data process first loads everything into `import_test`. Once loaded, it swaps the schemas.

But this means we are duplicating the import schema during the load, which is why we need too much extra allocation space. If we dropped the import schema and then loaded the extract CSV files straight in, we'd avoid duplicating anything and get that 'cushion' back.

We understand why it was done. Because of the decision to fetch returns data directly from the import schema when viewing them in the legacy UI, the previous team needed to add this precaution in case the import failed.

But as we are now moving to our new view return, which will be based solely on data in WRLS, this caution is no longer necessary.