AWS import performance tuning
https://eaflood.atlassian.net/browse/WATER-4855

The import process takes considerably longer in our AWS environments than it does locally. Specifically, it is the `src/modules/import-job/lib/licence-data-import.js` where the hold up occurs.

Locally, we see that the first 1000 licences take 2 mins 30 secs to complete. In our AWS `dev` environment which is running a [db.t3.medium](https://aws.amazon.com/rds/instance-types/) it takes 12 minutes.

In AWS, the CPU utilization is approx 75%, plenty of free memory, and read/write latency of 1ms or less. We seem to be well below both our baseline I/O operations/second (2000) and the max (11800) with reads around 10 and writes around 40 per second. Finally, the burst balance hovers around 99%.

So, this branch is all about trying to see what changes we can make to optimise the import for AWS (without _completely_ re-writing the legacy code!)

We start with getting a timing for the first 50 licenses with no changes, and then we'll start experimenting!