Fix logging errors when using omg()
https://eaflood.atlassian.net/browse/WATER-4855

With the import dealing with so many transactions, there are times we don't want to send every one to Errbit. It is enough that some indication will be registered in the email, and the actual error will be logged.

That is if it is logged!

We spotted during some testing tat an error was thrown, but what we saw in the logs was just an empty object.

```text
0|import  | [16:16:01.938] INFO (2407): import-job.licence-data-import: errored
0|import  |     licence: {
0|import  |       "ID": "21",
0|import  |       "LIC_NO": "6/33/03/*S/0033",
0|import  |       "AREP_SUC_CODE": "ARSUC",
0|import  |       "AREP_AREA_CODE": "ARCA",
0|import  |       "SUSP_FROM_BILLING": "N",
0|import  |       "AREP_LEAP_CODE": "C4",
0|import  |       "EXPIRY_DATE": "null",
0|import  |       "ORIG_EFF_DATE": "01/11/1966",
0|import  |       "ORIG_SIG_DATE": "null",
0|import  |       "ORIG_APP_NO": "null",
0|import  |       "ORIG_LIC_NO": "null",
0|import  |       "NOTES": "null",
0|import  |       "REV_DATE": "01/09/1972",
0|import  |       "LAPSED_DATE": "null",
0|import  |       "SUSP_FROM_RETURNS": "N",
0|import  |       "AREP_CAMS_CODE": "ACC01",
0|import  |       "X_REG_IND": "N",
0|import  |       "PREV_LIC_NO": "null",
0|import  |       "FOLL_LIC_NO": "null",
0|import  |       "AREP_EIUC_CODE": "ANOTH",
0|import  |       "FGAC_REGION_CODE": "1",
0|import  |       "SOURCE_CODE": "null",
0|import  |       "BATCH_RUN_DATE\r": "31/03/2025 22:22:00"
0|import  |     }
0|import  |     error: {}
```

Huh!

After some testing it seems to be how pino deals with the error instance. It works in `omfg()` because when we have an error instance, we add it to the log packet as `{ err: error }`. It seems to get pino to detail the error you have to add it to the log packet as `err:`.

So, this change updates how we pass in the data to `omg()` to cater for this.