AWS import performance tuning  (#1083)
https://eaflood.atlassian.net/browse/WATER-4855

The import process takes considerably longer in our AWS environments than locally. The holdup occurs in `src/modules/import-job/lib/licence-data-import.js`.

Locally, we see that the first 1000 licences take 2 minutes and 30 seconds to complete. In our AWS `dev` environment, which is running a [db.t3.medium](https://aws.amazon.com/rds/instance-types/), it takes 12 minutes.

In AWS, the CPU utilization is approximately 75%, plenty of free memory, and read/write latency is 1 ms or less. We seem to be well below both our baseline I/O operations per second (2000) and the max (11800), with reads around 10 and writes around 40 per second. Finally, the burst balance hovers around 99%.

We started this branch by testing what changes we could make to optimise the import for AWS without _completely_ rewriting the legacy code!

The things we tried in combination were

- Changing the number of configured PostgreSQL connections
- Running the individual licence processes sequentially instead of synchronously
- Processing licences concurrently (and how many)

We measured the timings for processing the first 50 licences, then when we had an idea of what, if anything, was working, the first 1000 licences.

The final changes we made were based on those results.

## First 50

| Connections | Concurrent licences | Concurrent processes | Time    |
|------------:|--------------------:|----------------------|--------:|
|          20 |                   0 | Yes                  | 1m 36s  |
|          40 |                   0 | Yes                  | 22s     |
|          20 |                   0 | No                   | 21s     |
|          40 |                   0 | No                   | 21s     |
|          20 |                  10 | No                   | 18s     |
|          40 |                  10 | No                   | 19s     |
|          40 |                  20 | No                   | 18s     |
|          40 |                  50 | No                   | 19s     |
|          40 |                  10 | Yes                  | 19s     |
|          40 |                   5 | No                   | 18s     |
|         100 |                   5 | No                   | 19s     |
|         100 |                 100 | No                   | 18s     |

## First 1000

| Connections | Concurrent licences | Concurrent processes | Time    |
|------------:|--------------------:|----------------------|--------:|
|         100 |                 100 | No                   | 4m 39s  |
|          20 |                  10 | No                   | 4m 7s   |
|          20 |                   5 | No                   | 3m 48s  |
|          20 |                   5 | Yes                  | 3m 42s  |
|          20 |                   5 | Yes                  | 3m 49s  |

## Final configuration

| Connections | Concurrent licences | Concurrent processes |
|------------:|--------------------:|----------------------|
|          20 |                   5 | No                   |